web_crawler/requirements.txt
python-dotenv==1.2.1
boto3==1.42.36
flask==3.1.2
flask-cors==6.0.2